refactor : 서비스로직 구현체로 분리 / feat : 파일 확장자 검증 추가 #8
